requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
crispy-bootstrap4==2022.1
cryptography==42.0.4
defusedxml==0.8.0rc2
Django==3.2
django-allauth==0.61.1
django-ckeditor==6.7.1
django-countries==7.5.1
django-crispy-forms==2.0
django-js-asset==2.2.0
django-rest-framework==0.1.0
django-tinymce==3.7.1
djangorestframework==3.14.0
google-api-core==2.17.1
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
msgpack==1.0.5
msgpack-python==0.5.6
oauthlib==3.2.2
packaging==23.2
Pillow==9.4.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-core==4.5.3
sqlparse==0.4.3
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.4.0
